requirements.txt
appdirs==1.4.4
asn1crypto==1.4.0
Brlapi==0.8.3
CacheControl==0.12.6
cffi==1.15.0
chardet==4.0.0
colorama==0.4.4
contextlib2==0.6.0.post1
cryptography==36.0.1
distlib==0.3.3
distro==1.6.0
html5lib==1.1
idna==3.3
libfdt==1.6.1
libvirt-python==7.8.0
louis==3.20.0
more-itertools==8.10.0
msgpack==1.0.3
ordered-set==4.0.2
packaging==20.9
pep517==0.12.0
ply==3.11
progress==1.6
pyalpm==0.10.6
pycairo==1.20.1
pycparser==2.21
PyGObject==3.42.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.0
pytoml==0.1.21
pyudev==0.22.0
requests==2.26.0
resolvelib==0.5.5
retrying==1.3.3
six==1.16.0
smbus==1.1
termcolor==1.1.0
toml==0.10.2
tomli==1.2.2
urllib3==1.26.7
webencodings==0.5.1
